requirements.txt
# Web框架与服务器
fastapi==0.95.2
uvicorn==0.22.0

# 数据库与ORM
sqlalchemy==1.4.49
mysql-connector-python==8.0.33
pymysql==1.0.3
pydantic==2.0.3

# 缓存
redis==4.5.4

# 环境变量管理
python-dotenv==1.0.0
envparse==0.2.0

# HTTP客户端
requests==2.31.0

# 爬虫相关
beautifulsoup4==4.12.2
feedparser==6.0.10
lxml==4.9.2
python-multipart==0.0.6

# 监控与日志
prometheus-client==0.17.0

# 数据处理
numpy==1.24.3
pandas==2.0.2
textblob==0.17.1

# 测试
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.1

# 工具库
humanize==4.6.0
dateparser==1.1.8

# 安全
hashing==0.2.3
oauthlib==3.2.2

# Vercel Edge Config
git+https://github.com/vercel/edge-config-python.git

# 其他
click==8.1.3
pyyaml==6.0